+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: `arrayToObject()` now takes "prop" function, emits uniform objects –
+  Where for non-object arrays the values of the returned object are now the
+  raw array items (Technically a breaking change.)
 
 ## 3.3.0